backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
alembic
PyJWT
python-multipart
bcrypt
passlib
chromadb
google-generativeai
requests
PyMuPDF
sentence-transformers
numpy
pandas
python-jose
fastapi-cors
aiofiles
